LA-224 Rename MJS extensions into CJS. Reason is that MJS refers to ESM while CJS refers to CommonJS. The latter is correct.
