+++ ArticleShapeExtractor/commands/CreateLeadArticle.idjs
@@ -1,6 +1,6 @@
 await (async function main() {
     const initBootstrap = require('../bootstrap.js');
-    const Container = require("../modules/Container.mjs");
+    const Container = require("../modules/Container.cjs");
 
     try {
         initBootstrap();
