+++ ArticleShapeExtractor/commands/ExtractArticleShapes.idjs
@@ -1,18 +1,18 @@
 await (async function main() {
     const { app } = require("indesign");
     const initBootstrap = require('../bootstrap.js');
-    const Container = require("../modules/Container.mjs");
+    const Container = require("../modules/Container.cjs");
     try {
         initBootstrap();
 
         if (app.documents.length === 0) {
-            const { NoDocumentOpenedError } = require('../modules/Errors.mjs');
+            const { NoDocumentOpenedError } = require('../modules/Errors.cjs');
             throw new NoDocumentOpenedError();
         }
 
         const doc = app.activeDocument;
         if (doc.articles.length === 0) {
-            const { NoArticlesInDocumentError } = require('../modules/Errors.mjs');
+            const { NoArticlesInDocumentError } = require('../modules/Errors.cjs');
             throw new NoArticlesInDocumentError();
         }
 
@@ -22,7 +22,7 @@ await (async function main() {
         const domains = require('uxp').storage.domains;
         const folder = await lfs.getFolder({initialDomain: domains.userDocuments});
         if (!folder || !folder.isFolder) {
-            const { NoFolderSelectedError } = require('../modules/Errors.mjs');
+            const { NoFolderSelectedError } = require('../modules/Errors.cjs');
             throw new NoFolderSelectedError();
         }
 
@@ -35,7 +35,7 @@ await (async function main() {
         // Export article shapes to disk.
         let exportCounter = await Container.resolve("ExportInDesignArticlesToFolder").run(doc, folder);
         if (exportCounter === 0) {
-            const { PrintLayoutAutomationError } = require('../modules/Errors.mjs');
+            const { PrintLayoutAutomationError } = require('../modules/Errors.cjs');
             throw new PrintLayoutAutomationError("No articles exported. Check if articles are properly named and have frames defined.");
         }
         alert(exportCounter + " articles have been exported as Article Shapes.");
