+++ ArticleShapeExtractor/generate.types.mjs
@@ -1,6 +1,6 @@
 /**
  * Understands how to (re)generate the types.d.js file.
- * For each mjs file the modules folder, its creates a typedef import.
+ * For each cjs file the modules folder, its creates a typedef import.
  */
 
 import fs from "fs";
@@ -10,10 +10,10 @@ const MODULES_DIR = path.join(process.cwd(), "modules");
 const OUTPUT_FILE = path.join(process.cwd(), "types.d.js");
 
 // Read module filenames
-const files = fs.readdirSync(MODULES_DIR).filter(f => f.endsWith(".mjs"));
+const files = fs.readdirSync(MODULES_DIR).filter(f => f.endsWith(".cjs"));
 
 if (files.length === 0) {
-    console.error("No .mjs files found in ./modules/");
+    console.error("No .cjs files found in ./modules/");
     process.exit(1);
 }
 
@@ -40,7 +40,7 @@ for (const file of files) {
 
 // Build final output.
 const output = `// IMPORTANT: DO NOT MANUALLY EDIT THIS FILE!
-// This file is auto generated by the generate-types.mjs tool.
+// This file is auto generated by the generate-types.cjs tool.
 // Simply run this tool when modules have changed.
 
 /**
