+++ ArticleShapeExtractor/modules/BrandSectionMapResolver.cjs
