+++ ArticleShapeExtractor/modules/BrandSectionResolver.cjs
