+++ ArticleShapeExtractor/modules/PageLayoutSettings.cjs
@@ -38,7 +38,7 @@ class PageLayoutSettings{
         app.scriptPreferences.measurementUnit = idd.MeasurementUnits.POINTS;
         try {
             if (doc.pages.length === 0) {
-                const { NoDocumentPagesError } = require('./Errors.mjs');
+                const { NoDocumentPagesError } = require('./Errors.cjs');
                 throw new NoDocumentPagesError();
             }
             for (let i = 0; i < doc.pages.length; i++) {
@@ -52,7 +52,7 @@ class PageLayoutSettings{
             await this.#saveOrComparePageLayoutSettings(settings, folder);
             exportedSuccessfully = true;
         } catch(error) {
-            const { ConfigurationError } = require('./Errors.mjs');
+            const { ConfigurationError } = require('./Errors.cjs');
             if (error instanceof ConfigurationError) {
                 this.#logger.error(error.message);
             } else {
@@ -149,7 +149,7 @@ class PageLayoutSettings{
             const settingsJson = JSON.stringify(settings, null, 4);
             const byteCount = await settingsFile.write(settingsJson, {format: formats.utf8}); 
             if (!byteCount ) {
-                const { ConfigurationError } = require('./Errors.mjs');
+                const { ConfigurationError } = require('./Errors.cjs');
                 const message = `Could not write into file "${manifestFoldername}/${settingsFilename}".\nPlease check access rights.`;
                 throw new ConfigurationError(message);
             }
@@ -157,7 +157,7 @@ class PageLayoutSettings{
             const settingsOfPrecedingLayout = JSON.parse(await settingsFile.read({format: formats.utf8}));
             const diff = this.#diffInDesignPageLayoutGrid(settings, settingsOfPrecedingLayout);
             if (diff != null) {
-                const { ConfigurationError } = require('./Errors.mjs');
+                const { ConfigurationError } = require('./Errors.cjs');
                 const message = "\n" 
                     + `A page setting for the current layout differs from the preceding layout, processed before.\n`
                     + `The '${diff.propertyPath}' setting for the current layout is '${diff.lhsValue}' but for the preceding layout is '${diff.rhsValue}'.\n`
@@ -179,7 +179,7 @@ class PageLayoutSettings{
         const pathsToCompare = [
             "columns.gutter",
             "baseline-grid.increment"
-            // Keep this list in sync with the diffInDesignPageLayoutGrid function in ArticleShapeUploader/modules/PageLayoutSettings.mjs
+            // Keep this list in sync with the diffInDesignPageLayoutGrid function in ArticleShapeUploader/modules/PageLayoutSettings.cjs
         ];
         for (const path of pathsToCompare) {
             const thisValue = this.#getPropertyValueByPath(lhsSettings, path);
