+++ ArticleShapeExtractor/modules/PreferencesManager.cjs
