+++ ArticleShapeExtractor/modules/VersionUtils.cjs
