+++ ArticleShapeExtractor/types.d.js
@@ -1,5 +1,5 @@
 // IMPORTANT: DO NOT MANUALLY EDIT THIS FILE!
-// This file is auto generated by the generate-types.mjs tool.
+// This file is auto generated by the generate-types.cjs tool.
 // Simply run this tool when modules have changed.
 
 /**
@@ -7,20 +7,20 @@
  * Cmd+Click on classes and functions makes its navigation feature work.
  */
 
-/** @typedef {import('./modules/BrandSectionMapResolver.mjs')} BrandSectionMapResolver */
-/** @typedef {import('./modules/BrandSectionResolver.mjs')} BrandSectionResolver */
-/** @typedef {import('./modules/Errors.mjs')} ArgumentError */
-/** @typedef {import('./modules/ExportInDesignArticlesToFolder.mjs')} ExportInDesignArticlesToFolder */
-/** @typedef {import('./modules/FileUtils.mjs')} FileUtils */
-/** @typedef {import('./modules/FitArticleWithAIService.mjs')} FitArticleWithAIService */
-/** @typedef {import('./modules/GenreResolver.mjs')} GenreResolver */
-/** @typedef {import('./modules/HttpLogger.mjs')} HttpLogger */
-/** @typedef {import('./modules/InDesignArticleService.mjs')} InDesignArticleService */
-/** @typedef {import('./modules/Logger.mjs')} Logger */
-/** @typedef {import('./modules/PageLayoutSettings.mjs')} PageLayoutSettings */
-/** @typedef {import('./modules/PlaService.mjs')} PlaService */
-/** @typedef {import('./modules/PreferencesManager.mjs')} PreferencesManager */
-/** @typedef {import('./modules/RegenerateArticleShapesService.mjs')} RegenerateArticleShapesService */
-/** @typedef {import('./modules/Settings.mjs')} Settings */
-/** @typedef {import('./modules/StudioJsonRpcClient.mjs')} StudioJsonRpcClient */
-/** @typedef {import('./modules/VersionUtils.mjs')} VersionUtils */
+/** @typedef {import('./modules/BrandSectionMapResolver.cjs')} BrandSectionMapResolver */
+/** @typedef {import('./modules/BrandSectionResolver.cjs')} BrandSectionResolver */
+/** @typedef {import('./modules/Errors.cjs')} ArgumentError */
+/** @typedef {import('./modules/ExportInDesignArticlesToFolder.cjs')} ExportInDesignArticlesToFolder */
+/** @typedef {import('./modules/FileUtils.cjs')} FileUtils */
+/** @typedef {import('./modules/FitArticleWithAIService.cjs')} FitArticleWithAIService */
+/** @typedef {import('./modules/GenreResolver.cjs')} GenreResolver */
+/** @typedef {import('./modules/HttpLogger.cjs')} HttpLogger */
+/** @typedef {import('./modules/InDesignArticleService.cjs')} InDesignArticleService */
+/** @typedef {import('./modules/Logger.cjs')} Logger */
+/** @typedef {import('./modules/PageLayoutSettings.cjs')} PageLayoutSettings */
+/** @typedef {import('./modules/PlaService.cjs')} PlaService */
+/** @typedef {import('./modules/PreferencesManager.cjs')} PreferencesManager */
+/** @typedef {import('./modules/RegenerateArticleShapesService.cjs')} RegenerateArticleShapesService */
+/** @typedef {import('./modules/Settings.cjs')} Settings */
+/** @typedef {import('./modules/StudioJsonRpcClient.cjs')} StudioJsonRpcClient */
+/** @typedef {import('./modules/VersionUtils.cjs')} VersionUtils */
